startlette-subscriptions/pyproject.toml
[tool.poetry]
name = "startlette-subscriptions"
version = "0.1.0"
description = ""
authors = ["Jonathan Kim <jkimbo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
strawberry-graphql = "^0.77.8"
starlette = "^0.16.0"
uvicorn = {extras = ["standard"], version = "^0.15.0"}

[tool.poetry.dev-dependencies]
black = "^21.9b0"
flake8 = "^3.9.2"
flake8-black = "^0.2.3"
flake8-bugbear = "^21.9.1"
mypy = "^0.910"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
